pongRebuild
You are to clone/download the provided project (Tic Tac Toe) and add an AI that will always pick the most optimal move. Your AI for selecting the computer's move should follow these rules (in order of importance):

a) If the computer can win immediately, it should do so

b) If the human's next move would result in a victory, the AI should choose that square to block

c) If the center square is open, choose it

d) If any of the corner squares are open, choose one of them

e) Choose any open square at random

Following these rules, if the AI goes first, it should never lose. It should always either win or tie. The player will only win if they go first and pick the exact optimal spaces.